+++ src/PixelEngine/Graphics/peglwidget.h
@@ -32,14 +32,14 @@ namespace PE
             ~PEGLWidget();
             World *GetWorld();
             void SetWorld(World *w);
-            double GetFPS();
+            float GetFPS();
 
         protected:
             void paintEvent(QPaintEvent *event) override;
-            double fps_sampling_rate = 2000;
+            float fps_sampling_rate = 2000;
             qint64 fps_start;
             int currentFPS = 0;
-            double lastFPS = 0;
+            float lastFPS = 0;
             QGLRenderer *renderer;
             World *world;
     };
