+++ src/PixelEngine/Graphics/peglwindow.cpp
@@ -39,7 +39,7 @@ void PEGLWindow::SetWorld(World *w)
     this->world = w;
 }
 
-double PEGLWindow::GetFPS()
+float PEGLWindow::GetFPS()
 {
     return this->lastFPS;
 }
@@ -58,7 +58,7 @@ void PEGLWindow::paintGL()
     qint64 current_time = QDateTime::currentDateTime().toMSecsSinceEpoch();
     if (current_time - this->fps_start > this->fps_sampling_rate)
     {
-        this->lastFPS = static_cast<double>(this->currentFPS) / (this->fps_sampling_rate / 1000);
+        this->lastFPS = static_cast<float>(this->currentFPS) / (this->fps_sampling_rate / 1000);
         this->currentFPS = 0;
         this->fps_start = current_time;
     }
